requirements.txt
praw
flask
flask-cors
numpy
pandas
python-dotenv
geopy
gunicorn